pyproject.toml
[project]
name = "kangabase"
version = "0.1.0"
description = "轻如 SQLite、易如 Supabase、为 Agent 而生的数据库框架"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    {name = "KangaBase Team", email = "team@kangabase.dev"}
]
keywords = ["database", "sqlite", "agent", "supabase", "contract"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
web = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.23",
    "jinja2>=3.1",
    "python-multipart>=0.0.6",
]

[project.scripts]
kangabase = "cli.main:cli"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["__init__", "agent"]

[tool.setuptools.packages.find]
include = ["core*", "cli*", "nl*", "web*", "migrations*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
kangabase = ["examples/**/*", "web/templates/**/*", "web/static/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
